Re-organize typed transaction file structure + add blob txns:
- Re-organize typed transactions as a directory with each `_TypedTransactionImplementation`
  as its own file.
- Add support for `BlobTransaction` (type = 3).
- Add relevant tests.

Also:

- Normalize transaction dict for ``TypedTransaction.as_dict()`` so that lists or tuples are converted to tuples for consistency and ease of comparison.
